Optimize withCompositionLocalProvider
Tip from Chuck for single values
