+++ circuit-foundation/src/commonMain/kotlin/com/slack/circuit/foundation/SaveableStateHolder.kt
@@ -9,6 +9,7 @@ import androidx.compose.runtime.saveable.LocalSaveableStateRegistry
 import androidx.compose.runtime.saveable.SaveableStateRegistry
 import androidx.compose.runtime.saveable.Saver
 import androidx.compose.runtime.saveable.rememberSaveable
+import com.slack.circuit.foundation.internal.withCompositionLocalProvider
 
 /**
  * This is a copy of [androidx.compose.runtime.saveable.SaveableStateHolder], tweaked so that
